requirements.txt
tika==2.6.0
tqdm==4.65.0
iscc-sdk==0.5.5